fix: Add queries element to AndroidManifest
This commit adds the <queries> element to the AndroidManifest.xml file
to address an issue where the app could not find a calendar app on
devices running Android 11 or higher. This is due to package visibility
changes introduced in Android 11. Adding the <queries> element with the
appropriate intent filter allows the app to discover and interact with
calendar apps.