Dev (#37)
* Fix

* Add rating feature

* fix: Handle timezones in session management and calendar export

This commit fixes a bug where the application did not correctly handle timezones when managing sessions and exporting them to the calendar. The code now uses the 'America/Toronto' timezone to ensure consistency across all users.

- Fixed timezone issue in  for adding events to the calendar, cancelling sessions, and loading sessions.
- Fixed timezone issue in  for loading available sessions.
- Fixed a lint error related to  in .
- Fixed a compilation error in .

* fix: Improve 'calendar not found' message

This commit improves the user experience by providing a more informative
message when a calendar app is not found on the device. Instead of a
generic 'No calendar app found' message, the user is now prompted to
install a calendar app to use the feature.

* fix: Add queries element to AndroidManifest

This commit adds the <queries> element to the AndroidManifest.xml file
to address an issue where the app could not find a calendar app on
devices running Android 11 or higher. This is due to package visibility
changes introduced in Android 11. Adding the <queries> element with the
appropriate intent filter allows the app to discover and interact with
calendar apps.

---------

Co-authored-by: onurqnel <onur.onel2230@gmail.com>
Co-authored-by: imanemm <imanemoussamarou@gmail.com>